(#2432) Add Actions workflow to run test script
Runs the test.bat/test.sh script on a new Github Actions workflow. Runs
the integration tests on Windows, the unit tests are already run by the
build.bat script. Runs only the unit tests on non-Windows as the
integration tests are not working there yet. The integration tests will
be run on other platforms with a later commit.

The test.yml workflow is scheduled to run at 10:47pm every day. The
specific minute is randomized instead of running on the hour to
reduce load.