(#2437) Fall back to CurrentUser scope when permission denied
When running on mono on a non-Windows system, the ProtectedData methods
use /usr/local/.mono/keypair as the folder to save keypairs. This
folder is not normally writable by non-root users, thus erroring when
used by a normal user.

This catches that error and falls back to using the CurrentUser scope.
Also, during data unprotection, it will try to fall back to CurrentUser
scope if the decryption fails for other reasons, so as to attempt to
decrypt data encrypted with CurrentUser if the LocalSystem scope starts
working.