(maint) Remove Coveralls environment variable
Moving this to be set within the AppVeyor UI, rather then in the
configuration file. I believe that the decryption of this variable
stopped working when the build was moved from under Rob's
AppVeyor account into the Chocolatey one, and this wasn't picked
up on, which meant that Coveralls was no longer being updated.