(#1332) Add switch to disable logging for archive extraction
Adds a switch "disableLogging" to Get-ChocolateyUnzip and
Install-ChocolateyZipPackage. This switch disable 7zip logging of
extracted filenames and does not write out the log of extracted
files. Install-ChocolateyZip passes the switch to Get-ChocolateyUnzip,
and the functionality is implemented into Get-ChocolateyUnzip.

This is useful for large archives with a very large number of files as
logging the extracted files can significantly slow down the extraction
process.