(#2450) Enable functions on non-Windows systems
This runs configTransformService and capture_package_files on
non-Windows platforms. Both appear to be fully functional, so there is
no reason not to run them.

They were added to non-Windows side of the if/else instead of being
moved from the Windows side so as to not screw up the functionality of
create_ignore_files_for_executables.