(2465) Fix numerous spelling mistakes
This was done using a GitHub Action which you can find details about
here: https://github.com/marketplace/actions/check-spelling

This may be something that we want to add into the group of GitHub
Actions that we have, but that is something for a different day. For
now, the fixing of the spelling mistakes is enough.

Signed-off-by: Josh Soref <jsoref@users.noreply.github.com>