(#2481) Only allow installing template packages on non-Windows
This adds a check for non-Windows systems that only allows package ids
ending in .template to install. This check is added to help provide
guidance to users, as the only package type currently supported on
non-Windows systems.

The check can be overridden if --force is specified. Because
upgrade_run calls install_run if the package is not already installed,
there is no need to duplicate this check in upgrade_run.

In the future, if the licensed extension becomes supported on
non-Windows systems, then the check will need to be amended to support
installing packages relating to licensed extension.