(#886) Add XML attributes for all arguments
This adds the remaining arguments in the install command's OptionSet
as elements to the packages.config xml serialization schema.

This is useful for users wanting arguments that were previously not
available. It is also a pre-requisite for exporting remembered
arguments, as some of the remembered arguments were not available
in the previous available schema.