(#449) Add template command
This allows a user to get information about all installed templates, or
a specific template. The list subcommand lists all installed templates.
The info subcommand lists verbose information about a specific template

Usage with -r should output a machine parsable list of the installed
templates. Usage with verbose (implied with the info subcommand)
outputs more information about the template(s).

The information displayed here can be gathered by running the list
command, config command and inspecting the contents of the
TemplatesLocation. However, this command allows the information to be
displayed in one location.