(#2434) Enable more scenarios, skip only shim and powershell facts
This enables more scenarios on non-windows platforms where the only
issues were that the facts relating to validating shimgen or powershell
script behavior did not work. These failures were expected as neither
shimgen or powershell run on non-windows platforms. So the individual
tests can be skipped, while enabling the rest of the scenario.