(#2481) Warn about unuspported packages on non-Windows
This adds a check for non-Windows systems that warns if the package id
does not end in .template on install. This check is added to help
provide guidance to users, as templates are the only package type
currently supported on non-Windows systems.

In the future, this check may be expanded to allow other packages
types, as needed.