(#747) Switch CryptoHashProvider to use FileStream
Previously, when getting the checksum of a file, this read the file
into a byte array to pass to ComputeHash. However, this was limited to
files of 2 GB or less, and was not efficent memory wise. This changes
the method to using a FileStream which is passed to ComputeHash, which
should allow arbitrary file sizes with better memory usage.

The unit test for this method had to be removed because the mock
FileSystem does not allow opening a FileStream, so instead, an
integration test has been created, which creates a large 2.5 GB file
which can then be used to ComputeHash on, to verify usage against larger
files. It was discussed whether this large file should be added to the
repository, but it was felt that the 15-20 seconds wait time to create
the file was a better solution.