(#747) Switch CryptoHashProvider to filestream for hashing files
Previously, when getting the checksum of a file, this read the file
into a byte array to pass to ComputeHash. However, this was limited to
files of 2gb or less, and was not efficent memory wise. This changes
the method to using a filestream which is passed to ComputeHash, which
should allow arbitrary file sizes with better memory usage.

The unit test for this method had to be removed because the mock
filesystem does not allow opening a file stream.