(maint) Add NUnitTestAdapter nuget packages
The NUnit test adapter extension does not work with VisualStudio 2022
and will not be updated to work it. The recommended alternative is to
use the nuget package instead of the extension. This PR adds the nuget
package to the packages.config for the test projects, so the tests can
be run directly in VisualStudio 2022.