(#2529) Reset configuration options before each test
Since there is a `new` ChocolateyConfiguration before each spec, it
invalidates the action that sets the configuration from the passed in
option because the old config is no longer used. Thus, the options set
should be reset so the help option can be re-added for each spec,
pointing to the new ChocolateyConfiguration object.