(#1896) Outdated with ignore-pinned use correct enhanced exit code
If there are no outdated unpinned packages, but there are are outdated
pinned packages, then an enhanced exit code of 2 will be incorrectly
used, while it should be an exit code of 0. This assumes enhanced exit
codes are enabled.

This corrects the check for if exit code 2 should be used to only
include unpinned packages, like the log output does.