(#1185) Add ability to run hook scripts
This adds the ability to run Powershell scripts (hook scripts) before
and after each type of package automation script runs. The hook scripts
can be for all packages, or for a specific package ID. They are looked
for recursively inside $env:ChocolateyInstall\hooks

The filename of the .ps1 file determines when they are run:
"<pre|post>-<install|beforemodify|uninstall>-<packageID|all>.ps1"

The first section is for timing of when the hook script runs.
Ore scripts run before the package automation script
(e.g. the ChocolateyInstall.ps1), while post scripts are run afterwards.
The second section of the filename determines what type of automation
script the hook is supposed to run for. The third section either
a string of the package ID the hook is to run for, or  the keyword all
to specify that it should be run for all package IDs.

If a hook needs to be run for more than one of these cases and/or it is
complex, the logic can be extracted out to another file or module and
dotsourced/imported.

If there are hook script(s) found, they will be run even if there is no
package automation script.

The hooks are run in the same context as package automation scripts
are, so they have access to the same environment variables as the
package scripts do.