(#1185) Add ability to handle .hook packages
This adds the ability for choco to install .hook packages that contain
hook PowerShell scripts, the same in concept as a .template or .extension
package.

For packages ending in ".hook", any files or directories inside a top
level folder called "hook" will be copied to a directory under
"$env:chocolatey\hooks\" named the same as the package ID with the
".hook" stripped off the end.

During upgrades, this directory will be removed, and updated files will
be copied over. During uninstalls, this directory will be removed.