(#2540) Python source disable all keyword on upgrade
This throws an exception if the all keyword is used on the python
alternate source for the upgrade command. On normal nuget sources, the
all keyword will upgrade all installed packages. However, on the python
source, it will try to upgrade a package named "all", which provides a
confusing error message for a user expecting all installed python
packages to be upgraded. This provides an explicit error message
explaining that the all keyword is not supported.

The other alternate sources do not require this, as they do not
implement upgrade at all.