(#2597) Add Pester tests to test the package built
This is a commit that is an amalgamation of 2 different projects with
minimal changes from those repositories.

The tests within the tests/chocolatey-tests directory comes
from the chocolatey-test-kitchen repository here:
https://gitlab.com/chocolatey/build-automation/chocolatey-test-kitchen/-/tree/11e0a76aa600cb1ea9d50b4b2a0f11f41c2ca751/

The helpers within the tests/helpers directory comes from here:
https://gitlab.com/chocolatey/build-automation/chocolatey-test-kitchen/-/tree/11e0a76aa600cb1ea9d50b4b2a0f11f41c2ca751/

Any package in the tests/packages directory comes from the
smoke-testing-packages repository here:
https://gitlab.com/chocolatey/build-automation/smoke-testing-packages/-/tree/d6945b7d324bc4fcfa4dd248fde47bb6b51a6500/

The only changes from those points in time are as follows:

* All of the tools directory in the mvcmusicstore directory has been
emptied except for a chocolateyInstall.ps1 file. These packages are only
used for choco info testing, and so we did not need these files. The
chocolateyInstall.ps1 has also been replaced with a single Write-Host
call.
* Get-chocolateyunzip-licensed package. The
cmake zip file in there has been replaced with an empty zip file as the
tests here are expected not to actually work, so we don't need the
content there.
* All ps1 files have been updated to ensure line endings are CRLF and
the file encoding is UTF8 with a BOM