(#2605) Pin to windows-2019 for GitHub Build image
The build and integration tests are failing on the windows-latest build
image due to a problem with ILMerge. Rather than trying to figure out
what is going on here, let's pin to windows-2019 instead, and we can
re-visit once we move over to using Chocolatey.Cake.Recipe for the
build.