(build) Remove AppVeyor build configuration
This is no longer being used really, since we have GitHub Action builds
running for Windows, Linux and Mac, and on top of that we have a new
TeamCity build running internally, to produce the signed artifacts for
publishing.  As such, having another CI build isn't required.  In
addition, currently the Chocolatey.Cake.Recipe doesn't have a build
provider for AppVeyor, which means that certain things won't work as
expected.