(#2462) Deprecate remote args for list command
This commit updates the list command to deprecate
the argument `--local-only` and arguments only valid
for remote sources when running the list command,
with a message that list will be made local only by default
in v2.0.0. The same message mentions to use search or find
instead for remote searches.

There are additionally several messages other places to ensure
that it will be difficult to miss the deprecation of remote
sources when calling the list command.