(#2630) Remove properties from ConfigFileSettings
These properties had previously been marked as Obsolete, and since we
are releasing the 2.0.0 version of Chocolatey, the decision has been
taken to remove these properties completely.

There were some usages of these properties, in the Equals and
GetHashCode methods for the class, but their removal isn't believed to
cause any issues.

In addition, we have moved away from using the custom HashCode class, in
favour of the new HashCode class that has been created in the .NET
Framework.  We didn't have access to this class before, since we were
still targetting .NET Framework 4.0.