(build) Added missing Kotlin import
While reviewing the various settings.kts files across the different
build configurations, it was believed that there was a duplicate Kotlin
import for xmlReport/XmlReport.  The duplicate was removed, however it
was found that this resulted in the build configuration not being
imported correctly.  A change was made to "fix" the casing of the
namespace to match with the name of the build feature, i.e. XmlReport.
Whlie this didn't result in any errors during import of the build
configuration, it did result in no tests being reported to TeamCity.

This change adds back in the "duplicate" import, with the mixed casing,
and sets the build feature name to be the lower case version.  This
basically takes it back to how it was before things were "optimized"!