(maint) Update deprecation notice for config settings
This commit updates the deprecation notice mentioned for
obsolete settings values which had initially been slated to
be removed in Chocolatey CLI v1.0.0. However, these are
required by CCM 0.7.0 and needs to be available until we
stop supporting this version.

As such, the deprecation notice is updated to reflect the
status, and the possibility of it being removed in v2.