(docs) Update committers docs with correct/fixed links
This commit updates the COMMITTERS.md file to use the
correct links to the documentation, as well as updating the
links that are pointing to pages that no longer exist to use
an archived version of the page.

The archive links should be updated to later posts that
are more up to date and still exists without having to rely
on the web archive, however in the interest of time and to
have valid links that users can use I opted to only point
to the web archive now.