(build) Additional tweaks to version numbers
After checking, and re-checking what happens on both an annotated and
non-annotated tags in the git repository, some additional changes were
needed to ensure that the generated version numbers matched what was
expected.

Comparions were made to the _BuildInfo.xml and SolutionVersion.cs file
from running the build_official.bat file on the 0.12.1 tag in the
chocolatey/choco repository, to what is generated in this repository
when running with an annotated tag, as well as a normal tag.  From what
can be seen, the corect version numbers are now being asserted, and
written into these files.