(#2656) Re-instate the update and version commands
This commit re-instates the ChocolateyVersionCommand and
ChocolateyUpdateCommand classes, however, it doesn't provide any
implementations for the methods within the ICommand interface.  In
addition, it also doesn't decorate the class with the CommandFor
attribute.

This has the effect that the commands still remain as unfound when
using Chocolatey CLI open source edition, however, it means that anyone
using an older version of the Chocolatey Licensed Extension (for
example perpetual licensed customers) will still have a basic
implementation of these commands that can be overridden.

The Exception that is thrown when one of the implemented methods is
executed is designed to match closely to the command not found error,
which is actually what we _want_ folks to see.

This change also has the side-effect of meaning that anyone directly
upgrading the chocolatey package, without first upgrading the
chocolatey.extension package, will no longer be broken after that
installation.