(#2663) Check file signature on 6 bytes
Turns out there's a thing in Windows 10+ where an otherwise empty file
was catalog signed. This means that all files that are BOM + \r\n end up
being marked as signed. This works around that so that Windows 10/11
systems get a profile update if they have an otherwise empty file.