(#2660) Add runtime check for compatible versions
In order to provide as much support to users of Chocolatey as possible,
we have added the concept of a minimum required Chocolatey version
attribute to the Chocolatey Licensed Assembly.  When reading this
assembly (if present), we will extract this attribute, and then use it
to make assertions about whether things are currently known to be
compatible.

When it is found that they are not compatible, a warning is shown
before and after the command execution to try to make it as clear as
possible that things might not work as expected.

A new feature flag has been introduced, disableCompatiblityChecks,
which a user can opt-in to, to prevent the warnings from showing, but
this is disabled by default.  The warning shows how this feature can be
turned on if they are sure that they know what they are doing.

In addition, a command line option has been added to allow by passing
these warnings without having to enable the feature, with the
skip-compatibility-checks option.

In addition, due to the fact that we can, and do, load the Chocolatey
Licensed Assembly from memory, the Location property can be null.  When
this is the case, and it hasn't been possible to extract the version
information, we reach into the custom attributes for the assembly to
pull out the information, as this is needed to assert on the
compatibility of the versions.