(#2660) Only show warnings when using license
When running choco.exe with no Chocolatey Licensed Extension in play the
compatibility warnings were still showing when they shouldn't. Make sure
to initialize the ChocolateyLicense with IsCompatible defaulting to true
until found to be anything else, as well as to only show the warnings
when it is known that a licensed version is being used.