(maint) Update regex for published date
The regex used for checking the information of packages was initially
based on the internally published package. This was published on a date
with 2 digits for the day. With the preparations made in this repo, the
date could be any day of the month. Updated the regex to ensure 1 or 2
digits for the day.