(#2400) Remove all tools from lib folder
Previously, the UppercuT build required that all tools were either
already installed on the machine, or present in the lib folder of the
repository.  This strategy is no longer required, as the
Chocolatey.Cake.Recipe already has the ability to download the tools
when required as part of the build process.