(#2400) Change nuspec folder structure
Chocolatey.Cake.Recipe is quite opinionated about the folder structure
that is used when creating Chocolatey and NuGet packages.  It expects
all nuspec files to be stored in a folder structure like the following:

- nuspec
  - chocolatey
  - nuget

With the UppercuT build, this structure existed, just in a slightly
different way.  This commit updates the folder structure to be what is
required.