(#2400) Update tests to work with Cake output
Cake outputs to a new file (and a lot to the terminal). This updates the
script used for running the tests to default to the new location.
We also update the Vagrantfile to silently absorb the output from the
build script as it seems to output too fast for vagrant to handle
consistently.