(#2400) Fix Docker build
In order for the Cake Build to correctly identify that it is running
on GitHub Actions, in order to assert teh correct branch name to use the
environment variables from the GitHub Actions build image have to be
passed through to the execution of Docker. We don't need all the
variables, just the ones used by Cake to make some assertions.