(#1956) Ignore versioned directories in Local Repository
This commit updates the ChocolateyLocalPackageRepository to allow a property
being set which allows our code to control whether any directories that are
versioned (side by side installed) to be ignored.

This is done to prevent NuGet Core throwing an error when it finds multiple
package versions of the same package being installed.