(#2742) store build output if build fails
When we build in the vagrant environment, we suppress the build output
as vagrant doesn't like all of the output the build sends at it (it
seems to come too fast). This stores that output into a file if the
build fails and then writes a message to that effect.