(#1185) Add tests for hooks
This adds tests for hook scripts. It validates that the skip-hooks
option is added, that hooks install/upgrade/uninstall correctly, and
that hooks are run at the correc times during
install/upgrade/uninstall.

As a bonus, this also adds a test case for "pure" portable packages,
that is packages without any package automation scripts. First, it
tests that they install correctly. Second it checks that hook scripts
will still run even if there is no chocolateyInstall.ps1

There are a few things added to the integration Scenario.cs to ensure
that hooks are correctly setup and reset for tests. The copying of the
hook scripts relies on the CommandName being set correctly.