(build) Switch to latest Chocolatey.Cake.Recipe
Turn off the execution of OpenCover when running the Chocolatey
integration tests.  FOr some reason, when executing OpenCover with
these tests, the standard output and error output don't seem to be
accessible.  Since some of the tests make assertions based on the
content of these outputs, the tests fail.

Since we don't _need_ the OpenCover output when running the integration
tests, let's turn off the execution, but leave it on when running the
unit tests.

The ability to have this granularity for when to run OpenCover was
became available in version 0.16.0 of Chocolatey.Cake.Recipe.