(#2771) Ensure documentation file is created
The NuGet package for chocolatey.lib expects that there is a
documentation file (chocolatey.xml) that needs to be included in the
package.  This file was being generated during a release build, but not
during a debug build.  As a result, the creation of the NuGet packages
was failing during a debug build, but happily working when doing a
release build.

This commit ensuring that the documentation file is correctly created
during a debug build.