(#2465) Revert correction to spelling mistake
This particular spelling mistake is on a public method, which can (and
is) used in some upstream projects.  As such, we would have to regard
this change as a breaking change, and it would have to go out as a
major release of Chocolatey.  This is not currently happening,
therefore we need to revert this change, and re-add this change in a
future release.