(maint) Update install/uninstall to new reset handler
This commit updates other handlers in the Nuget Service to use
the same approach when resetting configuration class as was
added when upgrading packages.

This allows us to use the same approach in the relevant code paths
without having to rely on manually deep copying the reference and
thus loosing the actual reference to the configuration object.