(#1185) Re-add old powershell run host method
Changing the interface or public methods are breaking changes which
we need to wait with implementing before the next major release.
As such, this commit adds back the old way of running the powershell
host by adding an overload with the original contract for the interface
that could possibly be used in other projects that makes use of the
Chocolatey.Lib library.

This overload has been made obsolete and will just call the new method
while setting any hook scripts as empty enumerables.