(#2693) Re-add old resolve or load method
Changing the interface or public methods are breaking changes which
we need to wait with implementing before the next major release.
As such, this commit adds back the old way of running the powershell
host by adding an overload with the original contract for the interface
that could possibly be used in other projects that makes use of the
Chocolatey.Lib library.