(#886) Update packages.config install logging
Update the debug logging when a packages.config file is being used to
output the start and end of the configuration on independent lines. This
allows us to update the test to get these start and end points and not
play guessing games with how long we think the output should be.
Updating this test allows Chocolatey Licensed Extension to be included
and the tests still pass.