(#2848) Removed `[A]ll - yes to all`
After some discussion with the Chocolatey team, it has been decided that
the prompt option for "All - yes to all" is rather verbose, and the need
to include "yes to all" is unneeded, and not in line with how other
package managers do it when prompted with similar choices.

This commit removes the "All - yes to all" and replaces it with simply
"All scripts", which is more succinct.  In addition, the unit/pester
tests have been updated to reflect this change.  It was necessary to
remove a unit test, since in the previous prompt "all - yes to all" was
actually treated as two options, split on the "-", i.e. both "all" and
"all - yes to all" were accepted answers to the prompt.  This is no
longer the case when the option is "All scripts", so only one unit test
is now required.