(#2591) Add headers to limit output commands
When a user asks for limited output from Chocolatey, it is not uncommon
to pipe that output to `ConvertFrom-String` or `ConvertFrom-Csv` and
manually add headers to get back an object. This allows for getting a
header row back so that the end user doesn't need to add their own
headers and discern what they are.

This also adds a StringResources static class that allows us to store
constant strings in and use them across the code to reduce duplication.

A new OutputHelpers class has been introduced in an attempt to simplify
the concatenation of the limited output, when a collection of strings
are joined to the "|" character.

Some new Pester tests have also been added to verify the output from the
commands, both when the new '--include-headers' option is used, as well
as the 'alwaysIncludeHeaders' feature.