(#2860) Add new PS helper to get maintainer specific paths
This commit adds a new PowerShell helper to Chocolatey CLI
that allows maintainers to acquire the paths that we have stored
as private/internal environment variables.

This allows maintainers to still make use of these variables through
this new helper while still retaining the ability to push the package
to Chocolatey Community Repository.