(build) Ensure ReleaseOfficial configuration used
When executing the build.official.sh file, the intention is clear that
the ReleaseOfficial configuration is being used. This is what is done in
the build.official.bat file, however, this was missed in the .sh file.

This came up as a problem when running builds on POSIX for the purposes
of doing the Docker image creations.