(build) Don't build NuGet packages on POSIX
In order to create a "working" NuGet package, we need to ILMerge
assemblies into the choco.exe and chocolatey.dll files.  This is currently
only possible when running on Windows. As such, ensure to set the
shouldRunNuGet parameter to false when not running on Windows. This
setting could have been made globally within Chocolatey.Cake.Recipe,
however, this would have prevented some projects needing to make the
change in the other direction, and this was seen as the lesser of the
two.