(#2867) Add --online option to all commands
This commit adds the --online option to all Chocolatey commands, to
allow opening of the online help documentation for the command that is
being executed.

For example, running:

choco install -h

will output the help documentation to the console, but running:

choco install -h --online

will not output anything to the command line, and instead open the
default browser to the help documentation for the command.