(#2870) Update sln and csproj files
Rather than use a "hand-crafted" update to the sln file to add the
ReleaseOfficialNo7zip configuration, let Visual Studio do it. This also
includes updates to the csproj files for the projects to make use of
the new configuraiton.

This new configuration was needed to allow the exclusion of the test
projects from the build, so that they are not included in the second
execution of MSBuild.