(build) Add parameter to not build AWS projects
Within Chocolatey.Cake.Recipe, the default is to attempt the publishing
of AWS Lamdba functions, when running on CI. This functionality was
added into Chocolatey.Cake.Recipe 0.17.0, and any project using this
version of recipe will attempt to do this. We may want to think about
changing this default, and instead, only enable it when we know that
there are AWS projects within the repository.

The reason that this change was needed was due to failing builds on the
GitHub Actions, when attempting to run on Ubuntu/Mac.

# Conflicts:
#	recipe.cake