(build) Add GitVersion.yml file
This is required on the support branch, since we need to be able to
control the creation of the "tag" which is asserted by GitVersion. By
default, for a support branch, this is an empty string, which doesn't
work in terms of how we assert the version number for use in generating
the various packages as part of the build.

Without this change in place, the asserted version number is invalid,
similar to 1.2.1--1036