(#2814) Alphabetically sort output of commands
In order to make the output of the choco list commands, for example
`choco feature list` as easy to consume as possible, the sort order
has been changed to alphabetical.  This should make it much easier to
find the current state, or value, or a feature, config, source, or
apikey.

Tests have been added where possible to ensure that the output orders
things as expected.