(build) Switch to running all tests
For the time being, while we work through the next scope of work, we
want to make sure that we are running _all_ tests. This will ensure
that we don't miss anything when bringing in some of the planned
breaking changes in the next release.

All tests were already running on GitHub Actions when doing the nightly
test build, however, this makes it so they run on every CI build.  This
change can be backed out, once we have shipped the next release.