(build) Stop publishing of code coverage metrics
Since we are now running all tests on each build, we have had to turn
off running of OpenCover, as there are known issues with this when
running the integration tests. As a result, there are no code coverage
metrics to be published, so let's simply comment this action step out
for now.