(build) Remove additional code coverage action
This really should have been in the previous commit, but it was
forgotten about.

For now, all code coverage action work needs to be ceased, since we are
not running OpenCover (the tool that actually generates the code
coverage results) since there is an issue when running this in
conjunction with the integration tests.