(build) Add overrides to ApplicationParameters
When running integration tests, it is necessary to control "where" the
unit test context believes things are located.  This was already done
for a number of properties that live on the ApplicationParameters
class, but others had not been done.

Notably, the HooksLocation property hadn't be overridden.  As a result,
when trying to run the integration tests in ReleaseOfficial mode, part
of the tests believed that the HooksLocation pointed at the
c:/programdata folder, where as another part believed it was located in
the current execution folder of the tests.  As a result, a number of
tests failed.

The opportunity has been taken to override all values on the
ApplicationParamaters class that may get to the same point when writing
integration tests, and also to change the order of where this is done
to match the order that the properties appear in the
ApplicationParameters class.