(build) Update to latest Chocolatey.Cake.Recipe
This includes the ability to control exactly what nuspec and nupkg
files are included in the main portions of the build process, via the
new globbing parameters.

This is required here, since we additionally create a new no7zip
package, which we don't want to publish anywhere, however, since it is
copied to the same output folder, it is getting scooped up
automatically, and published.