(maint) Corrected structure of solution folders
Before the introduction of Chocolatey.Cake.Recipe, the packaging files
for both the NuGet and Chocolatey packages for this project were stored
in a nuget folder.  This was changed to instead be a top level nuspec
folder, but the solution folder structure was never updated to reflect
this.

This commit addresses this, and also adds the VERIFICATION.txt file
into the solution folder.