(#2781) Add chocolateyPackageId to env variables
Previously the only env variable to reference a Chocolatey package id
was $env:ChocolateyPackageName. This is inconsistent with the nusepec
file where it is referenced as id.

This commit adds chocolateyPackageId and packageId to the available
environment variables available inside the PowerShell host allowing a
package maintainer to reference the id of a package with the more
idiomatic $env:ChocolateyPackageId.