(maint) Remove top level packages.config file
This file, which contained an additional source URL for a NuGet v2
source feed, is no longer required. Previously, this was necessary due
to the fact that we were attempting to restore the coveralls.io and
NUnit.Runners packages, which only existed on the NuGet v2 feed. These
packages are no longer required, therefore this file can be removed.