(#2738) Add net48 TFM folder in NuGet package
Now that we are building specifically for .NET 4.8, we should add the
generated chocolatey.dll file to the correct TFM folder within the
NuGet package, so that it is correctly located, and extracted when
consuming the NuGet package.

The Credits.txt file has been left at the root level of the lib folder,
rather than replicating into each TFM folder.