(maint) Use IO assertion instead of true/false assertion
This commit updates the assertions made to detect whether a
file or directory exists or not to instead of assertion on true/false
it instead asserts directly using NUnit v3 assertion classes.

This is done so if the test fails there will be more information
outputted of the failure than just telling whether something
was true or false.