(build) Update to 0.20.0 of Chocolatey.Cake.Recipe
This new version includes a fix for finding the correct version of
MSBuild that is required for a given build to succeed.  When we merged
in the work to uplift the Chocolatey Licensed project to use NUnit v3,
we included a bump in the langVersion for the projects, which
subsequently failed the build. This highlighted the fact that we
weren't using the version of MSBuild that we thought we were, mainly
due to the fact that we are running an older version of Cake within the
recipe. With this in mind, we have taken direct control over finding
exactly what MSBuild is required, and where it is located.